refactor: extract cycle/type boundaries into dedicated modules
