+++ src/components/customizer/armor/shared/LocationTypes.ts
@@ -1,0 +1,17 @@
+import { LocationPosition } from './MechSilhouette';
+
+export interface LocationArmorValues {
+  current: number;
+  maximum: number;
+  rear?: number;
+  rearMaximum?: number;
+}
+
+export interface LocationContentProps {
+  pos: LocationPosition;
+  data: LocationArmorValues;
+  showRear: boolean;
+  label: string;
+  isSelected: boolean;
+  isHovered: boolean;
+}
